refactor: move annotation schema to its own file
